Rename kb.c and kb.h to tau4.c and tau4.h
