+++ keyboards/tau4/keymaps/default/keymap.c
@@ -15,7 +15,7 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "kb.h"
+#include "tau4.h"
 #include "stdio.h"
 
 enum layers {
